src/dex-timeout.c: cast constant to 64 bits to avoid integer overflow
 from a coverity (static analyzer) report:
  src/dex-timeout.c:179:17:
    Type: Unintentional integer overflow (OVERFLOW_BEFORE_WIDEN)

  src/dex-timeout.c:179:17:
    overflow_before_widen: Potentially overflowing expression "1000000 * seconds" with type "int" (32 bits, signed) is evaluated using 32-bit arithmetic, and then used in a context that expects an expression of type "gint64" (64 bits, signed).
  src/dex-timeout.c:179:17:
    remediation: To avoid overflow, cast either "1000000" or "seconds" to type "gint64".

+++ src/dex-timeout.c
@@ -176,7 +176,7 @@ dex_timeout_new_deadline (gint64 deadline)
 DexFuture *
 dex_timeout_new_seconds (int seconds)
 {
-  gint64 usec = G_USEC_PER_SEC * seconds;
+  gint64 usec = (gint64) G_USEC_PER_SEC * seconds;
   return dex_timeout_new_deadline (g_get_monotonic_time () + usec);
 }
